Provisioning: Extract MigrationGuideNote; fix empty state for folders-only
- Move MigrationGuideNote into its own file with a unit test.
- Empty state now requires both no dashboards and no folders, so an
  instance with folders but no dashboards still renders the cards
  (review feedback).
- Drop statCardStyles/getSharedCardStyles — no longer shared; the value
  style is inlined into ResourceStatusCard.
