+++ public/app/features/provisioning/Migrate/Migrate.tsx
@@ -1,34 +1,14 @@
-import { css } from '@emotion/css';
 import { useMemo } from 'react';
 
-import { type GrafanaTheme2 } from '@grafana/data';
 import { t, Trans } from '@grafana/i18n';
-import { Alert, EmptyState, Spinner, Stack, Text, TextLink, useStyles2 } from '@grafana/ui';
+import { Alert, EmptyState, Spinner, Stack } from '@grafana/ui';
 import { getErrorMessage } from 'app/api/clients/provisioning/utils/httpUtils';
 import { useGetResourceStatsQuery } from 'app/api/clients/provisioning/v0alpha1';
 
-import { CONFIGURE_GRAFANA_DOCS_URL } from '../constants';
-
 import { MigrateToGitopsHeader } from './MigrateToGitopsHeader';
+import { MigrationGuideNote } from './MigrationGuideNote';
 import { OverviewStatCards } from './OverviewStatCards';
 import { aggregateFolderCounts, aggregateTotals, computeBreakdowns } from './stats';
-
-function MigrationGuideNote() {
-  const styles = useStyles2(getStyles);
-  return (
-    <div className={styles.intro}>
-      <Text color="secondary">
-        <Trans i18nKey="provisioning.migrate.intro">
-          The guided migration workflow is on its way. In the meantime, read the{' '}
-          <TextLink external href={CONFIGURE_GRAFANA_DOCS_URL}>
-            provisioning documentation
-          </TextLink>{' '}
-          to learn how Git Sync keeps Grafana and your repository in step.
-        </Trans>
-      </Text>
-    </div>
-  );
-}
 
 /**
  * Migrate to GitOps tab. Shows an overview of how much of the instance is
@@ -60,7 +40,7 @@ export function Migrate() {
     );
   }
 
-  if (totals.instanceTotal === 0) {
+  if (totals.instanceTotal === 0 && folderCounts.total === 0) {
     return (
       <Stack direction="column" gap={3}>
         <MigrateToGitopsHeader />
@@ -77,12 +57,3 @@ export function Migrate() {
     </Stack>
   );
 }
-
-const getStyles = (theme: GrafanaTheme2) => ({
-  intro: css({
-    padding: theme.spacing(2),
-    borderRadius: theme.shape.radius.default,
-    border: `1px solid ${theme.colors.border.weak}`,
-    background: theme.colors.background.secondary,
-  }),
-});
